# Use a couple of casts to speed up the MurmurString function