--- a/murmur_test.go
+++ b/murmur_test.go
@@ -10,6 +10,7 @@ import (
 
 var buf32 = make([]byte, 4)
 var buf64 = make([]byte, 8)
+var buf128 = make([]byte, 16)
 
 // Test that our abbreviated murmur hash works the same as upstream
 func TestMurmur(t *testing.T) {
@@ -30,6 +31,18 @@ func TestMurmur(t *testing.T) {
 		m := Murmur64(uint64(x))
 		if hash != m {
 			t.Errorf("Hash mismatch on 64 bit %d: expected 0x%X, got 0x%X\n", x, hash, m)
+		}
+	}
+
+	for i := 0; i < 100; i++ {
+		x := rand.Int63()
+		y := rand.Int63()
+		binary.LittleEndian.PutUint64(buf128, uint64(x))
+		binary.LittleEndian.PutUint64(buf128[8:], uint64(y))
+		hash := mmh3.Hash32(buf128)
+		m := Murmur128(uint64(x), uint64(y))
+		if hash != m {
+			t.Errorf("Hash mismatch on 128 bit %d,%d: expected 0x%X, got 0x%X\n", x, y, hash, m)
 		}
 	}
 }