The first cast saves a call to `runtime.stringtoslicebyte` which takes about 400 ns/op. The second cast converts the four bytes with about half the number of instructions, which saves about 100 ns/op. I've spent a bunch of time staring at the following documents, and I *think* this is safe (for our current platform and current version of Go).

https://golang.org/pkg/unsafe/#Pointer
https://golang.org/pkg/reflect/#StringHeader

The go runtime could still break this in the future, but I've attempted to document that.

Before:

goos: darwin
goarch: amd64
pkg: github.com/DataDog/hyperloglog
Benchmark100MurmerString-4 1000000	1798 ns/op	0 B/op	0 allocs/op

After:

goos: darwin
goarch: amd64
pkg: github.com/DataDog/hyperloglog
Benchmark100MurmerString-4 2000000 1207 ns/op	0 B/op	0 allocs/op